+++ src/passes/SimplifyGlobals.cpp
@@ -37,6 +37,7 @@
 #include <atomic>
 
 #include "ir/effects.h"
+#include "ir/linear-execution.h"
 #include "ir/properties.h"
 #include "ir/utils.h"
 #include "pass.h"
